chapter_9-1/requirements.txt
openai>=1.50.0
python-dotenv
